one component per file, clean code to conform to airbnb style guide
