+++ gulp/src/manageusers/status.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const Status = React.createClass({
+  render() {
+    let button = '';
+    if (this.props.status === true) {
+      button = <span className="label label-success">Active</span>;
+    }
+    else if (this.props.status === false) {
+      button = <span className="label label-warning">Pending</span>;
+    }
+    return (
+      <span>
+        {button}
+      </span>
+    );
+  },
+});
+
+export default Status;
