+++ gulp/src/manageusers/user.jsx
@@ -1,0 +1,299 @@
+/* global $ */
+
+import React from 'react';
+import Button from 'react-bootstrap/lib/Button';
+import HelpText from './helpText.jsx';
+import RolesMultiselect from './RolesMultiselect.jsx';
+import {Link} from 'react-router';
+import {validateEmail} from 'util/validateEmail';
+import {getCsrf} from 'util/cookie';
+
+const User = React.createClass({
+  propTypes: {
+    location: React.PropTypes.object.isRequired,
+    params: React.PropTypes.object.isRequired,
+    callUsersAPI: React.PropTypes.func,
+    history: React.PropTypes.object.isRequired,
+  },
+  getInitialState() {
+    // TODO Refactor to use basic Actions and the Dispatchers
+    return {
+      apiResponseHelp: '',
+      userEmail: '',
+      userEmailHelp: '',
+      roleMultiselectHelp: '',
+      availableRoles: [],
+      assignedRoles: [],
+      api_response_message: '',
+    };
+  },
+  componentDidMount() {
+    const action = this.props.location.query.action;
+
+    if (action === 'Edit') {
+      $.get('/manage-users/api/users/' + this.props.params.userId, function getUser(results) {
+        if (this.isMounted()) {
+          const userObject = results[this.props.params.userId];
+
+          const userEmail = userObject.email;
+
+          const availableRolesUnformatted = JSON.parse(userObject.roles.available);
+          const availableRoles = availableRolesUnformatted.map(function(obj) {
+            const role = {};
+            role.id = obj.pk;
+            role.name = obj.fields.name;
+            return role;
+          });
+
+          const assignedRolesUnformatted = JSON.parse(userObject.roles.assigned);
+          const assignedRoles = assignedRolesUnformatted.map(function(obj) {
+            const role = {};
+            role.id = obj.pk;
+            role.name = obj.fields.name;
+            return role;
+          });
+
+          this.setState({
+            userEmail: userEmail,
+            userEmailHelp: '',
+            roleMultiselectHelp: '',
+            apiResponseHelp: '',
+            availableRoles: availableRoles,
+            assignedRoles: assignedRoles,
+          });
+        }
+      }.bind(this));
+    } else {
+      $.get('/manage-users/api/roles/', function addUser(results) {
+        if (this.isMounted()) {
+          const availableRoles = [];
+          for (const roleId in results){
+            availableRoles.push(
+              {
+                'id': roleId,
+                'name': results[roleId].role.name,
+              }
+            );
+          }
+
+          this.setState({
+            userEmail: '',
+            userEmailHelp: '',
+            roleMultiselectHelp: '',
+            apiResponseHelp: '',
+            availableRoles: availableRoles,
+            assignedRoles: [],
+          });
+        }
+      }.bind(this));
+    }
+  },
+  onTextChange(event) {
+    this.state.userEmail = event.target.value;
+
+    const userEmail = this.state.userEmail;
+
+    if (validateEmail(userEmail) === false) {
+      this.setState({
+        userEmail: this.state.userEmail,
+        userEmailHelp: 'Invalid email',
+        availableRoles: this.refs.roles.state.availableRoles,
+        assignedRoles: this.refs.roles.state.assignedRoles,
+      });
+      return;
+    } else {
+      this.setState({
+        userEmail: this.state.userEmail,
+        userEmailHelp: '',
+        api_response_message: '',
+        availableRoles: this.refs.roles.state.availableRoles,
+        assignedRoles: this.refs.roles.state.assignedRoles,
+      });
+      return;
+    }
+  },
+  handleSaveUserClick() {
+    // Grab form fields and validate
+    // TODO: Warn user? If they remove a user from all roles, they will have to reinvite him.
+
+    const userId = this.props.params.userId;
+
+    let assignedRoles = this.refs.roles.state.assignedRoles;
+
+    const userEmail = this.state.userEmail;
+
+    if (validateEmail(userEmail) === false) {
+      this.setState({
+        userEmailHelp: 'Invalid email.',
+        roleMultiselectHelp: '',
+        availableRoles: this.refs.roles.state.availableRoles,
+        assignedRoles: this.refs.roles.state.assignedRoles,
+      });
+      return;
+    }
+
+    if (assignedRoles.length < 1){
+      this.setState({
+        userEmailHelp: '',
+        roleMultiselectHelp: 'Each user must be assigned to at least one role.',
+        availableRoles: this.refs.roles.state.availableRoles,
+        assignedRoles: this.refs.roles.state.assignedRoles,
+      });
+      return;
+    }
+
+    // No errors? Clear help text
+    this.setState({
+      availableRoles: this.refs.roles.state.availableRoles,
+      assignedRoles: this.refs.roles.state.assignedRoles,
+    });
+
+    // Format properly
+    assignedRoles = assignedRoles.map(function(obj) {
+       return obj.name;
+    });
+
+    // Determine URL based on action
+    const action = this.props.location.query.action;
+
+    let url = '';
+    if ( action === 'Edit' ) {
+      url = '/manage-users/api/users/edit/' + userId + '/';
+    } else {
+      url = '/manage-users/api/users/create/';
+    }
+
+    // Build data to send
+    const dataToSend = {};
+    dataToSend.csrfmiddlewaretoken = getCsrf();
+    dataToSend.assigned_roles = assignedRoles;
+    dataToSend.user_email = userEmail;
+
+    // Submit to server
+    $.post(url, dataToSend, function submitToServer(response) {
+      if ( response.success === 'true' ){
+        // Reload API
+        this.props.callUsersAPI();
+        // Redirect user
+        this.props.history.pushState(null, '/users');
+      }
+      else if ( response.success === 'false' ){
+        this.setState({
+          apiResponseHelp: response.message,
+          userEmail: this.state.userEmail,
+          availableRoles: this.refs.roles.state.availableRoles,
+          assignedRoles: this.refs.roles.state.assignedRoles,
+        });
+      }
+    }.bind(this))
+    .fail( function(xhr) {
+      if (xhr.status === 403){
+        this.setState({
+          apiResponseHelp: 'Unable to save user. Insufficient privileges.',
+        });
+      }
+    }.bind(this));
+  },
+  handleDeleteUserClick() {
+    const history = this.props.history;
+    // Temporary until I replace $.ajax jQuery with vanilla JS ES6 arrow function
+    const self = this;
+
+    if (confirm('Are you sure you want to delete this user?') === false) {
+      return;
+    }
+
+    const userId = this.props.params.userId;
+
+    const csrf = getCsrf();
+
+    // Submit to server
+    $.ajax( '/manage-users/api/users/delete/' + userId + '/', {
+      type: 'DELETE',
+      beforeSend: function(xhr) {
+        xhr.setRequestHeader('X-CSRFToken', csrf);
+      },
+      success: function( ) {
+        // Reload API
+        self.props.callUsersAPI();
+        // Redirect user
+        history.pushState(null, '/users');
+      }})
+      .fail( function(xhr) {
+        if (xhr.status === 403) {
+          this.setState({
+            apiResponseHelp: 'User not deleted. Insufficient privileges.',
+          });
+        }
+      }.bind(this));
+  },
+  render() {
+    let deleteUserButton = '';
+
+    let userEmailInput = '';
+
+    const action = this.props.location.query.action;
+
+    if (action === 'Edit') {
+      userEmailInput = <input id="id_userEmail" maxLength="255" name="id_userEmail" type="email" readOnly value={this.state.userEmail} size="35"/>;
+      deleteUserButton = <Button className="pull-right" onClick={this.handleDeleteUserClick}>Delete User</Button>;
+    } else {
+      userEmailInput = <input id="id_userEmail" maxLength="255" name="id_userEmail" type="email" value={this.state.userEmail} onChange={this.onTextChange} size="35"/>;
+    }
+
+    const userEmailHelp = this.state.userEmailHelp;
+    const roleMultiselectHelp = this.state.roleMultiselectHelp;
+    const apiResponseHelp = this.state.apiResponseHelp;
+
+    return (
+      <div>
+
+        <div className="row">
+          <div className="col-xs-12">
+            <div className="wrapper-header">
+              <h2>{action} User</h2>
+            </div>
+            <div className="product-card-full no-highlight">
+
+              <div className="row">
+                <div className="col-xs-12">
+                  <HelpText message={userEmailHelp} />
+                  <label htmlFor="id_userEmail">User Email*:</label>
+                  {userEmailInput}
+                </div>
+              </div>
+
+              <hr/>
+
+              <HelpText message={roleMultiselectHelp} />
+
+              <RolesMultiselect availableRoles={this.state.availableRoles} assignedRoles={this.state.assignedRoles} ref="roles"/>
+
+              <span id="role_select_help" className="help-text">To select multiple options on Windows, hold down the Ctrl key. On OS X, hold down the Command key.</span>
+
+              <hr />
+
+              <div className="row">
+
+                <div className="col-xs-12">
+                  <span className="primary pull-right">
+                    <HelpText message={apiResponseHelp} />
+                  </span>
+                </div>
+
+                <div className="col-xs-12">
+                  <Button className="primary pull-right" onClick={this.handleSaveUserClick}>Save User</Button>
+                  {deleteUserButton}
+                  <Link to="users" className="pull-right btn btn-default">Cancel</Link>
+                </div>
+              </div>
+
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  },
+});
+
+export default User;
